Run the Portable test subset standalone in CI
New `portable-tests` CI job runs this package's own @group Portable
tests with no demoshop, no live database, no Elasticsearch — verified
end-to-end from a real checkout of this repo, matrix on PHP 8.3/8.4,
204/204 green.

Same recipe as the sibling search-debug/search-ranking packages
(tests/_ci-standalone/): a direct TransferBusinessFactory call
generates Generated\Shared\Transfer\* into src/Generated/ (gitignored,
same convention any real Spryker project already applies to its own),
bypassing the full Zed Console/Kernel bootstrap and Locator entirely.

tests/_ci-standalone/Generated/Shared/Search/PageIndexMap.php is the
same checked-in, point-in-time snapshot the sibling packages carry —
this package's own OptimizationRunner-adjacent tests construct a
document containing PageIndexMap::SCORES, a constant this package
cannot regenerate alone. See the file's own docblock for the full
reasoning.

Two real tag corrections found while verifying against the live suite,
not the crude grep heuristic: AutomatedWeightOptimizationControllerTest
and SaturationPointCalibrationControllerTest both hit a real Propel
connection through their own Controller->Facade->EntityManager chain
(their own docblocks already say "INTEGRATION TEST" and describe
building a real database row) — now NeedsDatabase, not Portable.

phpcs/phpmd (all three rulesets)/rector all needed a src/Generated/
exclusion (freshly generated build output, gitignored) and phpcs/rector
needed one more for tests/_ci-standalone/Generated/ (the checked-in
core-generated fixture, which must stay byte-identical to core's own
output).

Also fixes a pre-existing, unrelated RecastingRemovalRector finding in
RankEvalRunner.php/ParameterVectorMapper.php (two redundant (float)
casts on expressions already typed float) — surfaced by the same
rector-version-drift pattern found earlier this session in the sibling
search-ranking package, fixed the same way rather than left blocking
the rector CI job.

Part of the cross-repo test-suite portability plan (W3, all three
phases now built).

+++ src/SprykerCommunity/Zed/SearchRankingOptimizer/Business/Optimization/ParameterVectorMapper.php
@@ -373,7 +373,7 @@ class ParameterVectorMapper implements ParameterVectorMapperInterface
         }
 
         foreach ($this->getFreeSpecificityKeys() as $key) {
-            $vector[] = (float)match ($key) {
+            $vector[] = match ($key) {
                 'specificityCurveExponent' => $configurationTransfer->getSpecificityCurveExponent() ?? 1.0,
                 'specificityWeightExponent' => $configurationTransfer->getSpecificityWeightExponent() ?? 1.0,
                 'specificityWeightShiftMagnitude' => $configurationTransfer->getSpecificityWeightShiftMagnitude() ?? 0.0,
